+++ packages/schema/src/lts/index.ts
@@ -4,3 +4,4 @@ export * as v1_6_0 from "./1.6.0";
 export * as v1_7_0 from "./1.7.0";
 export * as v1_8_0 from "./1.8.0";
 export * as v1_9_0 from "./1.9.0";
+export * as v1_10_0 from "./1.10.0";