+++ packages/examples/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @versaprotocol/examples
 
+## 1.10.0
+
+### Minor Changes
+
+- Update to schema_version 1.10.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @versaprotocol/schema@1.10.0
+
 ## 1.9.0
 
 ### Minor Changes